// src/logistics-api/model/updateTruckByLogistician.ts
/**
 * Logistics API
 * Logistics API provides services to find and manage entities for logistics
 *
 * OpenAPI spec version: 1.0
 * Contact: vladislavprokopenko1@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Object representing Truck updating by Logistician
 */
export interface UpdateTruckByLogistician { 
    model?: string;
    /**
     * Truck capacity, tons
     */
    capacity?: number;
    /**
     * Total operating time, years
     */
    totalOperatingTime?: number;
    technicalCondition?: UpdateTruckByLogistician.TechnicalConditionEnum;
    currentCity?: string;
    currentState?: string;
}
export namespace UpdateTruckByLogistician {
    export type TechnicalConditionEnum = 'OK' | 'NOK';
    export const TechnicalConditionEnum = {
        OK: 'OK' as TechnicalConditionEnum,
        NOK: 'NOK' as TechnicalConditionEnum
    };
}